// types.d.ts
type LeadingZero = `0${number}`;
type MM = LeadingZero | 10 | 11 | 12;
type DD = LeadingZero | `${1 | 2}${number}` | 30 | 31;
export type FormatDateString = `${number}-${MM}-${DD}`;
export type MonthsCount = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12;
export type Positions = 'bottom' | 'top' | 'center' | 'left' | 'right';
export type YPosition = 'bottom' | 'top';
export type XPosition = 'left' | 'center' | 'right';
export type PositionToInput = 'auto' | XPosition | [YPosition, XPosition];
export type Range<N extends number, Acc extends number[] = []> = Acc['length'] extends N ? Acc[number] : Range<N, [...Acc, Acc['length']]>;
// ...
export type CalendarSystem = 'gregory' | 'islamic-umalqura';
export type CalendarDateParts = {
    year: number;
    month: Range<12>;
    day: number;
};